refactor: handle thread tabs in ThreadViewModel
